requirements.txt
appdirs==1.4.3
diff-match-patch==20121119
Django==1.11
django-debug-toolbar==1.8
django-extensions==1.7.9
django-filter==1.0.4
django-import-export==0.5.1
djangorestframework==3.5.3
et-xmlfile==1.0.1
jdcal==1.3
node==0.9.17
odfpy==1.3.5
odict==1.5.2
openpyxl==2.4.8
plumber==1.3.1
pytz==2017.2
PyYAML==3.12
requests==2.12.1
six==1.10.0
sqlparse==0.2.3
tablib==0.11.5
unicodecsv==0.14.1
update==0.4.4
virtualenv==15.1.0
xlrd==1.0.0
xlwt==1.2.0
zope.component==4.3.0
zope.deprecation==4.2.0
zope.event==4.2.0
zope.interface==4.3.3
zope.lifecycleevent==4.1.0
